+++ src/util.cpp
@@ -160,17 +160,17 @@ vector<string> _directions{"+", "-", "<", ">"};
 // Mapping between stone types and vector representations of those stone types.
 // The stones are always represented from the perspective of the active player.
 // {0, 0, 0} == empty
-map<char, vector<int>> player_stone_type_to_vec{
-    {'E', {0, 0, 0}},
-    {'F', {0, 0, 1}},
-    {'S', {0, 1, 0}},
-    {'C', {0, 0, 1}},
+map<char, vector<float>> player_stone_type_to_vec{
+    {'E', {0.0, 0.0, 0.0}},
+    {'F', {0.0, 0.0, 1.0}},
+    {'S', {0.0, 1.0, 0.0}},
+    {'C', {0.0, 0.0, 1.0}},
 };
-map<char, vector<int>> opponent_stone_type_to_vec{
-    {'E', {0, 0, 0}},
-    {'F', {1, 1, 0}},
-    {'S', {1, 0, 1}},
-    {'C', {1, 1, 0}},
+map<char, vector<float>> opponent_stone_type_to_vec{
+    {'E', {0.0, 0.0, 0.0}},
+    {'F', {1.0, 1.0, 0.0}},
+    {'S', {1.0, 0.0, 1.0}},
+    {'C', {1.0, 1.0, 0.0}},
 };
 
 /*************************************
